favorite feature
fixed post structure

+++ api/app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\FilterPostRequest;
 use App\Http\Requests\PostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use App\Models\Photo;
 use App\Models\Post;
 use App\Models\Profile;
 use Illuminate\Support\Facades\Auth;
@@ -22,19 +23,28 @@ class PostController extends Controller
 
         $postData = $request->validated();
         $postData["profile_id"] = $profile_id;
-        unset($postData["photos"]);
-
+        unset($postData["outside_photos"],$postData["inside_photos"]);
         $uploadedPhotoPaths = [];
         DB::beginTransaction();
 
         try {
-            $post = Post::create($postData);
+            $post = Post::query()->create($postData);
 
-            if ($request->hasFile("photos")) {
-                $uploadedPhotoPaths = $this->storePhotosToPost(
+            if ($request->hasFile("outside_photos")) {
+                $outsidePhotoPaths = $this->storePhotosToPost(
                     $post,
-                    $request->file("photos"),
+                    $request->file("outside_photos"),
+                    Photo::TYPE_OUTSIDE
                 );
+                $uploadedPhotoPaths = array_merge($uploadedPhotoPaths, $outsidePhotoPaths);
+            }
+            if ($request->hasFile("inside_photos")) {
+                $insidePhotoPaths = $this->storePhotosToPost(
+                    $post,
+                    $request->file("inside_photos"),
+                    Photo::TYPE_INSIDE
+                );
+                $uploadedPhotoPaths = array_merge($uploadedPhotoPaths, $insidePhotoPaths);
             }
 
             DB::commit();
@@ -42,7 +52,9 @@ class PostController extends Controller
             return response()->json(
                 [
                     "message" => "Post created successfully",
-                    "post" => $post->load("photos"),
+                    "post" => $post->load(
+                        "outsidePhotos",
+                    "insidePhotos"),
                 ],
                 201,
             );
@@ -67,7 +79,6 @@ class PostController extends Controller
     }
 
 
-
     public function update(UpdatePostRequest $request, $PostId)
     {
         $user_id = Auth::user()->id;
@@ -80,19 +91,24 @@ class PostController extends Controller
         DB::beginTransaction();
 
         try {
-            $post->update($request->except("photos"));
+            $post->update($request->except(['outside_photos', 'inside_photos']));
 
-            if ($request->hasFile("photos")) {
-                $this->deleteOldPhotos($post);
-                $this->storePhotosToPost($post, $request->file("photos"));
+
+            if ($request->hasFile("outside_photos")) {
+                $this->deleteOldPhotos($post,Photo::TYPE_OUTSIDE);
+                $this->storePhotosToPost($post, $request->file("outside_photos"),Photo::TYPE_OUTSIDE);
             }
-
+            if ($request->hasFile("inside_photos")) {
+                $this->deleteOldPhotos($post,Photo::TYPE_INSIDE);
+                $this->storePhotosToPost($post, $request->file("inside_photos"),Photo::TYPE_INSIDE);
+            }
             DB::commit();
 
             return response()->json(
                 [
                     "message" => "Post updated successfully",
-                    "post" => $post->load("photos"),
+                    "post" => $post->load("outsidePhotos"
+                        ,"insidePhotos"),
                 ],
                 200,
             );
@@ -109,23 +125,40 @@ class PostController extends Controller
         }
     }
 
-    private function storePhotosToPost(Post $post, array $photos)
+    private function storePhotosToPost(Post $post, array $photos,String $type)
     {
+        $uploadedPaths = [];
+
         foreach ($photos as $photoFile) {
-            $path = $photoFile->store("post_photos");
+            $path = $photoFile->store("post_photos/{$post->id}/$type");
 
             $post->photos()->create([
+                "post_id"=>$post->id,
                 "file_path" => $path,
+                "type"=>$type
             ]);
+            $uploadedPaths[] = $path;
         }
+        return $uploadedPaths;
     }
 
-    private function deleteOldPhotos(Post $post)
+    private function deleteOldPhotos(Post $post,String $type)
     {
-        foreach ($post->photos as $photo) {
-            Storage::delete($photo->file_path);
+        if($type == Photo::TYPE_OUTSIDE){
+
+            foreach ($post->outsidePhotos as $photo) {
+                Storage::delete($photo->file_path);
+            }
+            $post->outsidePhotos()->delete();
         }
-        $post->photos()->delete();
+
+        if($type == Photo::TYPE_INSIDE){
+            foreach ($post->insidePhotos as $photo) {
+                Storage::delete($photo->file_path);
+            }
+
+            $post->insidePhotos()->delete();
+        }
     }
 
     public function getHomepageFeed()
@@ -133,11 +166,7 @@ class PostController extends Controller
 
 
         $postQ = Post::query()
-            ->with([
-                "photos" => function ($query) {
-                    $query->orderBy("created_at", "desc");
-                },
-            ])
+            ->with(['outsidePhotos','insidePhotos'])
             ->latest();
         $user = Auth::guard('sanctum')->user();
         if ($user) {
@@ -164,7 +193,7 @@ class PostController extends Controller
     {
         $details=Post::query()
             ->with('profile')
-            ->with('photos')
+            ->with(['outsidePhotos','insidePhotos'])
             ->withAvg('ratings' , 'rating')
             ->withCount('ratings')
             ->with(['ratings'=>function($query){
@@ -182,11 +211,7 @@ class PostController extends Controller
     public function getUsersPosts($ProfileId)
     {
         $posts = Post::query()
-            ->with([
-                "photos" => function ($query) {
-                    $query->orderBy("created_at", "desc");
-                },
-            ])
+            ->with(['outsidePhotos','insidePhotos'])
             ->where("profile_id", $ProfileId)
             ->latest()
             ->paginate(20);
@@ -203,11 +228,7 @@ class PostController extends Controller
 
         $posts = $profile
             ->posts()
-            ->with([
-                "photos" => function ($query) {
-                    $query->orderBy("created_at", "desc");
-                },
-            ])
+            ->with(['outsidePhotos','insidePhotos'])
             ->latest()
             ->paginate(20);
 
@@ -251,13 +272,13 @@ class PostController extends Controller
             $profile=$user->profile;
             $profile_id = $profile->id;
 
-            $posts=$query->with('Photos')
+            $posts=$query->with(['outsidePhotos','insidePhotos'])
                 ->whereNotIn('profile_id', [$profile_id])
                 ->paginate(20);
 
             return response()->json($posts,200);
         }
-        $posts=$query->with('Photos')
+        $posts=$query->with(['outsidePhotos','insidePhotos'])
         ->paginate(20);
 
         return response()->json($posts,200);
